address

+++ main.h
@@ -10,7 +10,7 @@ int handle_id(long j);
 int handle_c(int c);
 int handle_s(char *j);
 int custom_specifier(char *c);
-int handle(unsigned long b, int c, int a);
+int handle(unsigned int b, int c, int a);
 void dbuff(int a, char *b);
 int handle_p(void * p);
 #endif
